Unified NOC test
