+++ test/ttexalens/unit_tests/test_base.py
@@ -14,7 +14,7 @@ def init_default_test_context():
         return tt_exalens_init.init_ttexalens()
 
 
-def init_test_context(use_noc1: bool = True):
+def init_test_context(use_noc1: bool = False):
     if use_noc1:
         assert not os.getenv("TTEXALENS_TESTS_REMOTE"), "Remote testing for NOC1 not supported"
         return tt_exalens_init.init_ttexalens(use_noc1=True)
